refactor: common을 global로 패키지명 변경
