+++ backend/src/main/java/com/zzang/chongdae/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.common.config;
+package com.zzang.chongdae.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
