+++ backend/src/main/java/com/zzang/chongdae/offering/repository/entity/OfferingEntity.java
@@ -1,6 +1,6 @@
 package com.zzang.chongdae.offering.repository.entity;
 
-import com.zzang.chongdae.common.repository.entity.BaseTimeEntity;
+import com.zzang.chongdae.global.repository.entity.BaseTimeEntity;
 import com.zzang.chongdae.member.repository.entity.MemberEntity;
 import com.zzang.chongdae.offering.domain.OfferingMeeting;
 import com.zzang.chongdae.offering.domain.OfferingPrice;
